requirements.txt
openai~=0.25.0
argparse~=1.4.0
python-dotenv~=0.21.0
